backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
cryptography>=42.0
boto3>=1.34
drf-yasg>=1.21.7
djangorestframework-simplejwt>=5.3.1
sentry-sdk>=2.0.0
django-celery-results>=2.5.1
python-dateutil>=2.9.0
gunicorn>=21.2.0
